+++ src/Facades/DatabaseBackup.php
@@ -3,10 +3,11 @@
 namespace Mhmdomer\DatabaseBackup\Facades;
 
 use Illuminate\Support\Facades\Facade;
+use Mhmdomer\DatabaseBackup\DatabaseBackup as DatabaseBackupDatabaseBackup;
 
 /**
  * @method static array getBackupFiles()
- * @method static string getLatestBackupFile()
+ * @method static ?string getLatestBackupFile()
  *
  * @see \Mhmdomer\DatabaseBackup\DatabaseBackup
  */
@@ -14,6 +15,6 @@ class DatabaseBackup extends Facade
 {
     protected static function getFacadeAccessor()
     {
-        return DatabaseBackup::class;
+        return DatabaseBackupDatabaseBackup::class;
     }
 }